Minor code inspectation changes

+++ src/main/java/org/fit/cssbox/layout/TableCellBox.java
@@ -331,7 +331,7 @@ public class TableCellBox extends BlockBox
             if (width.isPercentage())
             {
             	wrelative = true;
-            	percent = (int) Math.round(width.getValue());
+            	percent = Math.round(width.getValue());
                 if (percent == 0)
                     wrelative = false; //consider 0% as absolute 0
             }
